+++ CHANGES.md
@@ -3,6 +3,8 @@
 UNRELEASED
 * Four new fields (inactive) added to Solr: nsfw_probability,is_nsfw,is_virus And virus_description. Fields are inactive in default solrwayback bundle.
 * Navigation tracker auto poll increased from 2 seconds to 60 seconds.
+* Upgraded Solr dependencies to 9.10.1
+* Switched from deprecated NoOpResponseParser to the new InputStreamResponseParser. Caching of SolrResponses is now handled by storing the response in the NetarchiveSolrClient.
 
 5.4.2
 -----